Change count variable type from int to size_t in print_list function

+++ singly_linked_lists/0-print_list.c
@@ -1,7 +1,5 @@
 #include <stdio.h>
 #include "lists.h"
-#include <stddef.h>
-#include <stdlib.h>
 /**
  * print_list - prints all the elements of a list_t list
  * @h: pointer to the head of the list
@@ -10,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h != NULL)
 	{
